refactor: rename `[project]` to `[workspace]` in manifest (#2510)
### Description

This is the first of a few PRs that will introduce the concept of
`workspace`s and `package`s.

With `pixi build` we want to have a distinction between:

* `pixi.toml` as an environment manager (which it currently is), and
* `pixi.toml` as a package build description.

These are two different use cases that we currently mix. Some keys are
related to package building e.g. `[host-dependencies]`,
`[build-dependencies]`, `[build]`. These only make sense in the context
of building a package out of a project. When we create an environment we
currently also just throw all the different dependency types on a big
pile and dont really make a distinction.

So we decided to split the `pixi.toml` into two distinct parts (in the
same file).

* `workspace`: Renamed from `project`. If a `pixi.toml` contains a
`workspace` it will manage environments.
* `package`: If this key is present the package can be build into a
package. It does **not** manage any environments.

Both `[workspace]` and `[package]` can be defined in the same file. If
the `[workspace]` is missing a "default" one is implicitly used.

### This PR

This PR only moves some files around in `pixi_manifest` and renamed
`[project]` to `[workspace]`. `[project]` is still allowed but will be
deprecated later.
