+++ crates/pixi_manifest/src/pyproject.rs
@@ -13,7 +13,7 @@ use toml_edit::DocumentMut;
 
 use super::{
     error::{RequirementConversionError, TomlError},
-    DependencyOverwriteBehavior, Feature, ParsedManifest, SpecType,
+    DependencyOverwriteBehavior, Feature, SpecType, WorkspaceManifest,
 };
 use crate::{error::DependencyError, FeatureName};
 
@@ -26,7 +26,7 @@ pub struct PyProjectManifest {
 
 #[derive(Deserialize, Debug, Clone)]
 pub struct Tool {
-    pub pixi: Option<ParsedManifest>,
+    pub pixi: Option<WorkspaceManifest>,
     pub poetry: Option<ToolPoetry>,
 }
 
@@ -90,7 +90,7 @@ impl PyProjectManifest {
         self.tool().and_then(|t| t.poetry.as_ref())
     }
 
-    fn pixi(&self) -> Option<&ParsedManifest> {
+    fn pixi(&self) -> Option<&WorkspaceManifest> {
         self.tool().and_then(|t| t.pixi.as_ref())
     }
 
@@ -105,7 +105,7 @@ impl PyProjectManifest {
     ///  - the `[project]` table
     ///  - the `[tool.poetry]` table
     pub fn name(&self) -> Option<String> {
-        if let Some(pixi_name) = self.pixi().and_then(|p| p.project.name.as_ref()) {
+        if let Some(pixi_name) = self.pixi().and_then(|p| p.workspace.name.as_ref()) {
             return Some(pixi_name.clone());
         }
         if let Some(pyproject) = &self.project {
@@ -122,7 +122,7 @@ impl PyProjectManifest {
     ///  - the `[project]` table
     ///  - the `[tool.poetry]` table
     fn description(&self) -> Option<String> {
-        if let Some(pixi_description) = self.pixi().and_then(|p| p.project.description.as_ref()) {
+        if let Some(pixi_description) = self.pixi().and_then(|p| p.workspace.description.as_ref()) {
             return Some(pixi_description.to_string());
         }
         if let Some(pyproject_description) =
@@ -141,7 +141,7 @@ impl PyProjectManifest {
     ///  - the `[project]` table
     ///  - the `[tool.poetry]` table
     fn version(&self) -> Option<String> {
-        if let Some(pixi_version) = self.pixi().and_then(|p| p.project.version.as_ref()) {
+        if let Some(pixi_version) = self.pixi().and_then(|p| p.workspace.version.as_ref()) {
             return Some(pixi_version.to_string());
         }
         if let Some(pyproject_version) = self.project.as_ref().and_then(|p| p.version.as_ref()) {
@@ -158,7 +158,7 @@ impl PyProjectManifest {
     ///  - the `[project]` table
     ///  - the `[tool.poetry]` table
     fn authors(&self) -> Option<Vec<String>> {
-        if let Some(pixi_authors) = self.pixi().and_then(|p| p.project.authors.as_ref()) {
+        if let Some(pixi_authors) = self.pixi().and_then(|p| p.workspace.authors.as_ref()) {
             return Some(pixi_authors.clone());
         }
         if let Some(pyproject_authors) = self.project.as_ref().and_then(|p| p.authors.as_ref()) {
@@ -201,8 +201,8 @@ impl PyProjectManifest {
     ///  - one environment is created per group with the same name
     ///  - each environment includes the feature of the same name
     ///  - it will also include other features inferred from any self references
-    ///    to other groups of optional dependencies (but won't for dependency groups,
-    ///    as recursion between groups is resolved upstream)
+    ///    to other groups of optional dependencies (but won't for dependency
+    ///    groups, as recursion between groups is resolved upstream)
     pub fn environments_from_extras(&self) -> Result<HashMap<String, Vec<String>>, Pep735Error> {
         let mut environments = HashMap::new();
         if let Some(extras) = self.optional_dependencies() {
@@ -225,7 +225,8 @@ impl PyProjectManifest {
         if let Some(groups) = self.dependency_groups().transpose()? {
             for group in groups.into_keys() {
                 let normalised = group.replace('_', "-");
-                // Nothing to do if a group of optional dependencies has the same name as the dependency group
+                // Nothing to do if a group of optional dependencies has the same name as the
+                // dependency group
                 if !environments.contains_key(&normalised) {
                     environments.insert(normalised.clone(), vec![normalised]);
                 }
@@ -244,7 +245,7 @@ pub enum PyProjectToManifestError {
     DependencyGroupError(#[from] Pep735Error),
 }
 
-impl TryFrom<PyProjectManifest> for ParsedManifest {
+impl TryFrom<PyProjectManifest> for WorkspaceManifest {
     type Error = PyProjectToManifestError;
 
     fn try_from(item: PyProjectManifest) -> Result<Self, Self::Error> {
@@ -257,10 +258,10 @@ impl TryFrom<PyProjectManifest> for ParsedManifest {
         // Set pixi project name, version, description and authors (if they are not set)
         // with the ones from the `[project]` or `[tool.poetry]` tables of the
         // `pyproject.toml`.
-        manifest.project.name = item.name();
-        manifest.project.description = item.description();
-        manifest.project.version = item.version().and_then(|v| v.parse().ok());
-        manifest.project.authors = item.authors();
+        manifest.workspace.name = item.name();
+        manifest.workspace.description = item.description();
+        manifest.workspace.version = item.version().and_then(|v| v.parse().ok());
+        manifest.workspace.authors = item.authors();
 
         // TODO:  would be nice to add license, license-file, readme, homepage,
         // repository, documentation, regarding the above, the types are a bit
@@ -314,7 +315,8 @@ impl TryFrom<PyProjectManifest> for ParsedManifest {
 
         // For each group of optional dependency or dependency group,
         // create a feature of the same name if it does not exist,
-        // and add pypi dependencies, filtering out self-references in optional dependencies
+        // and add pypi dependencies, filtering out self-references in optional
+        // dependencies
         let project_name = item.package_name();
         for (group, reqs) in groups {
             let feature_name = FeatureName::Named(group.to_string());
@@ -371,8 +373,7 @@ mod tests {
     use rattler_conda_types::{ParseStrictness, VersionSpec};
 
     use crate::{
-        manifests::manifest::Manifest, pypi::PyPiPackageName, DependencyOverwriteBehavior,
-        FeatureName,
+        manifests::Manifest, pypi::PyPiPackageName, DependencyOverwriteBehavior, FeatureName,
     };
 
     const PYPROJECT_FULL: &str = r#"
