+++ src/project/environment.rs
@@ -88,7 +88,7 @@ impl<'p> Environment<'p> {
             .solve_group
             .map(|solve_group_idx| SolveGroup {
                 project: self.project,
-                solve_group: &self.project.manifest.parsed.solve_groups[solve_group_idx],
+                solve_group: &self.project.manifest.workspace.solve_groups[solve_group_idx],
             })
     }
 
@@ -308,7 +308,7 @@ impl<'p> HasFeaturesIter<'p> for Environment<'p> {
         let manifest = self.manifest();
         let environment_features = self.environment.features.iter().map(|feature_name| {
             manifest
-                .parsed
+                .workspace
                 .features
                 .get(&FeatureName::Named(feature_name.clone()))
                 .expect("feature usage should have been validated upfront")
