+++ tests/data/non-satisfiability/binary-spec-source-record/pixi.lock
