+++ tests/data/non-satisfiability/source-dependency/child-package/pixi.toml
