+++ tests/data/non-satisfiability/source-dependency/pixi.toml
