+++ tests/data/satisfiability/source-dependency/child-package/pixi.toml
